VC Server/requirement.txt
1.Conda:4.5.12
2.Django framework:2.1.7